+++ packages/vscode-extension/esbuild.config.mjs
@@ -47,6 +47,7 @@ const external = [
   "serialport",
   "sharp",
   "tree-sitter-wasms",
+  "undici",
   "web-tree-sitter",
   "word-extractor",
 ];